implement callback resets and DiscreteCallbacks in SSAStepper
Right now it requires that the user knows to call `reset_aggregated_jumps!`. This can be fixed by hoisting this call all the way up to DiffEqBase and into `apply_discrete_callback!`, which is a bit of work so I'm punting on that for now but will open an issue.