Fix reset_aggregated_jumps! which is currently a no-op
Before this commit a call to `reset_aggregated_jumps!(integrator)` has no effect at all. Note that the jump aggregator is stored in the `integrator.cb` field and not in `integrator.opts.callback` (the latter is used to store additional user-provided callbacks). I suppose this commit fixes the code to do what it was intended to do.