Modify solve!/step! logic for proper tstops handling
This ensures that tstops are visited by the integrator, even when no more
jumps will happen before the end of the solution interval.

This change also slightly modifies the effect of DiffEqBase.step!. The
SSAIntegrator will refuse to step further if it is at the end of the
integration interval.