Fix out of bounds indexing for SSAStepper's __init
In SSAStepper's DiffEqBase.__init function, when the
kwarg save_start=false a local variable u gets initialized as an empty array.
Later in that function u[1] is accessed leading to an error.
This is fixed in this commit.

Additionally make sizehint! more accurate in the edge case where
save_everystep=false and isempty(saveat). This is
particularly helpful in situations where one
frequently remakes JumpProblems and has to avoid allocations for every
DiffEqBase.solve call.