Added ExtendedJumpArray dispatch for ODE_DEFAULT_NORM
This change reduces the number of allocations (to zero!) required
to take the norm of an ExtendedJumpArray. Previously, ODE_DEFAULT_NORM
fell back to a default, allocating implementation.

Added tests to check that this norm definition is equivalent to the
old one, in addition to an explicit check that the new norm is not
allocating.