Add non-default broadcasting of ExtendedJumpArray's.
The previous ExtendedJumpArray code contains code that implements part
of the broadcastable interface, but is not actually used due to the
entire interface not being included. This commit replaces that broadcast
code, reusing much of the arg packing/repacking.

Because of this, the current code uses the fallback indexing method when
broadcasting, which requires branching inside the broadcast loop. This
branch is likely well-branch-predicted, but the presence of the branch
prevents kernel functions from being fused into efficent SIMD
instructions.

This defines a ExtendedJumpArrayStyle, a broadcast style that includes
two sub-styles that represent the styles for broadcasting the `u` and
`jump_u` entries. `copyto!` overloads are specified that do the efficent
thing and call `copyto!`, using the broadcast computation tree, on `u`
and `jump_u` separately. Binary style rules are defined such that
broadcasts fall back to the old method if you do something like adding
an ExtendedJumpArray to a normal vector.

A Julia >=1.9 extension is added for FastBroadcast.jl, where the
equivilant of the `copyto!` call, `fast_materialize!` is similarlly
overloaded.